cleanup: Remove duplicate and obsolete files
- Delete 3 redundant test coverage boost files
- Remove 3 .broken test files (replaced by working versions)
- Delete 3 legacy/backup API client files (4,500+ lines of duplicates)
- Remove obsolete create_dummy_data.py from root
- Consolidate test_client_fixed.py → test_client.py (17 tests passing)
- Move documentation files to my_docs/ folder

Repository now clean with:
- No duplicate code
- Clear file organization
- All functionality preserved
- Scripts folder reserved for dev/demo utilities only

Files removed: 11 total
Lines of code removed: ~4,500+ duplicate lines
