--- tests/test_coverage_boost.py
@@ -1,167 +1,0 @@
-"""Focused integration tests to boost coverage quickly."""
-
-import os
-import sys
-from unittest.mock import Mock, patch
-
-import pytest
-
-# Add src to path
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "..", "src"))
-
-from ticker_converter.api_clients.client import AlphaVantageClient
-from ticker_converter.api_clients.constants import AlphaVantageFunction, OutputSize
-from ticker_converter.api_clients.exceptions import AlphaVantageAPIError
-
-
-class TestCoverageBoost:
-    """Focused tests to quickly boost coverage above 40%."""
-
-    def test_api_client_initialization(self):
-        """Test basic API client functionality."""
-        # Test initialization
-        client = AlphaVantageClient()
-        assert client is not None
-
-        # Test with custom key
-        client_with_key = AlphaVantageClient(api_key="test_key")
-        assert client_with_key.api_key == "test_key"
-
-    def test_constants_basic_functionality(self):
-        """Test that constants work as expected."""
-        # Test OutputSize
-        assert OutputSize.COMPACT.value == "compact"
-        assert OutputSize.FULL.value == "full"
-
-        # Test Functions
-        assert AlphaVantageFunction.TIME_SERIES_DAILY.value == "TIME_SERIES_DAILY"
-        assert AlphaVantageFunction.OVERVIEW.value == "OVERVIEW"
-
-    def test_exception_hierarchy(self):
-        """Test exception functionality."""
-        # Test base exception
-        error = AlphaVantageAPIError("Test error")
-        assert str(error) == "Test error"
-        assert isinstance(error, Exception)
-
-        # Test with error code
-        error_with_code = AlphaVantageAPIError("Test error", "ERR_001")
-        assert error_with_code.error_code == "ERR_001"
-
-    @patch("src.ticker_converter.api_clients.client.requests.get")
-    def test_api_client_make_request_basic(self, mock_get):
-        """Test basic request functionality."""
-        # Setup mock response
-        mock_response = Mock()
-        mock_response.status_code = 200
-        mock_response.json.return_value = {"test": "data"}
-        mock_response.raise_for_status.return_value = None
-        mock_get.return_value = mock_response
-
-        client = AlphaVantageClient(api_key="test_key")
-
-        # Test basic request
-        result = client._make_request({"function": "TIME_SERIES_DAILY"})
-        assert result == {"test": "data"}
-
-    def test_config_loading(self):
-        """Test configuration loading."""
-        from ticker_converter.api_clients.client import get_api_config
-
-        config = get_api_config()
-
-        assert config is not None
-        assert hasattr(config, "api_key")
-        assert hasattr(config, "base_url")
-        assert hasattr(config, "timeout")
-
-    def test_data_processors_basic(self):
-        """Test basic data processing functionality."""
-        from ticker_converter.api_clients.data_processors import (
-            convert_time_series_to_dataframe,
-        )
-
-        # Test with simple data
-        time_series = {
-            "2025-08-14": {
-                "1. open": "100.00",
-                "2. high": "105.00",
-                "3. low": "99.00",
-                "4. close": "103.00",
-                "5. volume": "1000000",
-            }
-        }
-
-        result = convert_time_series_to_dataframe(time_series)
-        assert result is not None
-        assert len(result) > 0
-
-    def test_utils_basic(self):
-        """Test basic utility functions."""
-        from ticker_converter.api_clients.utils import calculate_backoff_delay
-
-        # Test backoff calculation
-        delay = calculate_backoff_delay(0)
-        assert delay >= 12
-
-        delay = calculate_backoff_delay(1)
-        assert delay >= 24
-
-    @patch("src.ticker_converter.data_ingestion.nyse_fetcher.AlphaVantageClient")
-    def test_nyse_fetcher_basic(self, mock_client_class):
-        """Test basic NYSE fetcher functionality."""
-        from ticker_converter.data_ingestion.nyse_fetcher import NYSEDataFetcher
-
-        # Setup mock
-        mock_client = Mock()
-        mock_client_class.return_value = mock_client
-
-        fetcher = NYSEDataFetcher()
-
-        # Test basic properties
-        assert fetcher.magnificent_seven_symbols is not None
-        assert len(fetcher.magnificent_seven_symbols) == 7
-        assert "AAPL" in fetcher.magnificent_seven_symbols
-
-    def test_api_models_basic(self):
-        """Test basic API models functionality."""
-        from ticker_converter.data_models.api_models import StockPrice
-
-        # Test model creation
-        stock_price = StockPrice(
-            symbol="AAPL",
-            date="2025-08-14",
-            open=100.0,
-            high=105.0,
-            low=99.0,
-            close=103.0,
-            volume=1000000,
-        )
-
-        assert stock_price.symbol == "AAPL"
-        assert stock_price.open == 100.0
-
-    def test_market_data_basic(self):
-        """Test basic market data functionality."""
-        from ticker_converter.data_models.market_data import MarketDataManager
-
-        # Test initialization
-        manager = MarketDataManager()
-        assert manager is not None
-
-    @patch("src.ticker_converter.cli.main")
-    def test_cli_basic(self, mock_main):
-        """Test basic CLI functionality."""
-        from ticker_converter import cli
-
-        # Test that CLI module loads
-        assert cli is not None
-        assert hasattr(cli, "main")
-
-    def test_config_basic(self):
-        """Test basic configuration functionality."""
-        from ticker_converter.config import DatabaseConfig
-
-        # Test config creation
-        config = DatabaseConfig()
-        assert config is not None
