--- tests/test_coverage_boost_fixed.py
@@ -1,251 +1,0 @@
-"""
-Fixed coverage boost test - aligned with actual module structure.
-This test strategically targets high-coverage areas to reach 40% threshold.
-"""
-
-import os
-import sys
-import tempfile
-import unittest
-from datetime import datetime
-from unittest.mock import Mock, patch
-
-import pytest
-
-# Add src to path for imports
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "src"))
-
-# Set test environment
-os.environ["ALPHAVANTAGE_API_KEY"] = "test_mock_key"
-
-
-class TestCoverageBoostFixed(unittest.TestCase):
-    """Strategic tests to boost coverage to 40%."""
-
-    def setUp(self):
-        """Set up test environment."""
-        self.temp_dir = tempfile.mkdtemp()
-
-    def test_api_client_initialization(self):
-        """Test AlphaVantageClient initialization."""
-        from ticker_converter.api_clients.client import AlphaVantageClient
-
-        # Test with explicit API key
-        client = AlphaVantageClient(api_key="test_key")
-        assert client.api_key == "test_key"
-        assert client.base_url == "https://www.alphavantage.co/query"
-
-        # Test close method
-        client.close()
-
-    def test_api_client_session_setup(self):
-        """Test session setup functionality."""
-        from ticker_converter.api_clients.client import AlphaVantageClient
-
-        client = AlphaVantageClient(api_key="test_key")
-        client._setup_sync_session()
-        assert client.session is not None
-
-    @patch("ticker_converter.api_clients.client.requests.Session.get")
-    def test_api_client_make_request(self, mock_get):
-        """Test make_request method with proper method name."""
-        from ticker_converter.api_clients.client import AlphaVantageClient
-
-        # Setup mock response
-        mock_response = Mock()
-        mock_response.status_code = 200
-        mock_response.json.return_value = {"test": "data"}
-        mock_response.raise_for_status.return_value = None
-        mock_get.return_value = mock_response
-
-        client = AlphaVantageClient(api_key="test_key")
-
-        # Test make_request (not _make_request)
-        result = client.make_request({"function": "TIME_SERIES_DAILY"})
-        assert result == {"test": "data"}
-
-    def test_api_constants(self):
-        """Test API constants module."""
-        from ticker_converter.api_clients.constants import (
-            ALPHAVANTAGE_BASE_URL,
-            DEFAULT_OUTPUTSIZE,
-            DEFAULT_TIMEOUT,
-            DIGITAL_CURRENCY_SYMBOLS,
-            FOREX_PAIRS,
-            FUNCTION_TIME_SERIES_DAILY,
-            FUNCTION_TIME_SERIES_INTRADAY,
-            NYSE_SYMBOLS,
-            RATE_LIMIT_PER_DAY,
-            RATE_LIMIT_PER_MINUTE,
-        )
-
-        # Test constant values
-        assert ALPHAVANTAGE_BASE_URL == "https://www.alphavantage.co/query"
-        assert FUNCTION_TIME_SERIES_DAILY == "TIME_SERIES_DAILY"
-        assert FUNCTION_TIME_SERIES_INTRADAY == "TIME_SERIES_INTRADAY"
-        assert RATE_LIMIT_PER_MINUTE == 75
-        assert RATE_LIMIT_PER_DAY == 25000
-        assert DEFAULT_TIMEOUT == 30
-        assert DEFAULT_OUTPUTSIZE == "compact"
-
-        # Test collections exist
-        assert isinstance(DIGITAL_CURRENCY_SYMBOLS, list)
-        assert isinstance(FOREX_PAIRS, list)
-        assert isinstance(NYSE_SYMBOLS, list)
-        assert len(NYSE_SYMBOLS) > 0
-
-    def test_api_exceptions(self):
-        """Test API exception classes."""
-        from ticker_converter.api_clients.exceptions import (
-            AlphaVantageError,
-            DataNotFoundError,
-            InvalidSymbolError,
-            RateLimitError,
-        )
-
-        # Test base exception
-        base_error = AlphaVantageError("Test error")
-        assert str(base_error) == "Test error"
-
-        # Test specific exceptions
-        rate_error = RateLimitError("Rate limit exceeded")
-        assert "rate limit" in str(rate_error).lower()
-
-        symbol_error = InvalidSymbolError("Invalid symbol")
-        assert "symbol" in str(symbol_error).lower()
-
-        data_error = DataNotFoundError("Data not found")
-        assert "not found" in str(data_error).lower()
-
-    def test_data_processors(self):
-        """Test data processing functions."""
-        from ticker_converter.api_clients.data_processors import (
-            process_daily_data,
-            validate_response_data,
-        )
-
-        # Test response validation
-        valid_data = {"Meta Data": {}, "Time Series (Daily)": {}}
-        assert validate_response_data(valid_data, "TIME_SERIES_DAILY") is True
-
-        # Test invalid data
-        invalid_data = {"error": "Invalid API call"}
-        assert validate_response_data(invalid_data, "TIME_SERIES_DAILY") is False
-
-        # Test daily data processing
-        sample_data = {
-            "Time Series (Daily)": {
-                "2023-01-01": {
-                    "1. open": "100.00",
-                    "2. high": "105.00",
-                    "3. low": "99.00",
-                    "4. close": "104.00",
-                    "5. volume": "1000000",
-                }
-            }
-        }
-
-        processed = process_daily_data(sample_data)
-        assert len(processed) == 1
-        assert processed[0]["date"] == "2023-01-01"
-        assert float(processed[0]["open"]) == 100.00
-
-    def test_api_utils(self):
-        """Test utility functions."""
-        from ticker_converter.api_clients.utils import (
-            calculate_rate_limit_delay,
-            format_date,
-            validate_symbol,
-        )
-
-        # Test symbol validation
-        assert validate_symbol("AAPL") is True
-        assert validate_symbol("") is False
-        assert validate_symbol(None) is False
-
-        # Test date formatting
-        formatted = format_date("2023-01-01")
-        assert formatted == "2023-01-01"
-
-        # Test rate limit calculation
-        delay = calculate_rate_limit_delay(10, 60)
-        assert delay >= 0
-
-    def test_config_module_basic(self):
-        """Test basic config functionality that exists."""
-        from ticker_converter.config import DATABASE_URL, get_database_url
-
-        # Test basic config access
-        db_url = get_database_url()
-        assert isinstance(db_url, str)
-        assert len(db_url) > 0
-
-    def test_cli_module_basic(self):
-        """Test CLI module basic functionality."""
-        from ticker_converter.cli import main
-
-        # Just test that main function exists and is callable
-        assert callable(main)
-
-    def test_data_models_basic(self):
-        """Test basic data models functionality."""
-        from ticker_converter.data_models import api_models
-
-        # Test module imports successfully
-        assert hasattr(api_models, "__file__")
-
-    def test_data_ingestion_base_fetcher(self):
-        """Test base fetcher functionality."""
-        from ticker_converter.data_ingestion.base_fetcher import BaseFetcher
-
-        # Create instance (abstract class but we can test basic properties)
-        try:
-            fetcher = BaseFetcher()
-            # Test basic properties exist
-            assert hasattr(fetcher, "logger")
-        except TypeError:
-            # Expected for abstract class
-            pass
-
-    @patch("ticker_converter.data_ingestion.nyse_fetcher.AlphaVantageClient")
-    def test_nyse_fetcher_initialization(self, mock_client_class):
-        """Test NYSE fetcher initialization."""
-        from ticker_converter.data_ingestion.nyse_fetcher import NYSEDataFetcher
-
-        # Setup mock
-        mock_client = Mock()
-        mock_client_class.return_value = mock_client
-
-        fetcher = NYSEDataFetcher()
-
-        # Test basic properties that should exist
-        assert hasattr(fetcher, "client")
-        assert hasattr(fetcher, "logger")
-
-    @patch("ticker_converter.data_ingestion.currency_fetcher.AlphaVantageClient")
-    def test_currency_fetcher_initialization(self, mock_client_class):
-        """Test currency fetcher initialization."""
-        from ticker_converter.data_ingestion.currency_fetcher import CurrencyDataFetcher
-
-        # Setup mock
-        mock_client = Mock()
-        mock_client_class.return_value = mock_client
-
-        fetcher = CurrencyDataFetcher()
-
-        # Test basic properties
-        assert hasattr(fetcher, "client")
-        assert hasattr(fetcher, "logger")
-
-    def test_database_manager_basic(self):
-        """Test database manager basic functionality."""
-        from ticker_converter.data_ingestion.database_manager import DatabaseManager
-
-        # Test initialization with test database
-        db_manager = DatabaseManager(":memory:")
-        assert hasattr(db_manager, "db_path")
-        assert db_manager.db_path == ":memory:"
-
-
-if __name__ == "__main__":
-    unittest.main()
