--- tests/test_minimal_coverage_boost.py
@@ -1,229 +1,0 @@
-"""
-Minimal coverage boost test - only tests actual existing functions/classes.
-Strategic targeting to reach 40% coverage quickly.
-"""
-
-import os
-import sys
-import unittest
-from unittest.mock import Mock, patch
-
-# Add src to path for imports
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "src"))
-
-# Set test environment
-os.environ["ALPHAVANTAGE_API_KEY"] = "test_mock_key"
-
-
-class TestMinimalCoverageBoost(unittest.TestCase):
-    """Minimal tests targeting actual existing code."""
-
-    def test_api_client_initialization_and_methods(self):
-        """Test AlphaVantageClient initialization and basic methods."""
-        from ticker_converter.api_clients.client import AlphaVantageClient
-
-        # Test initialization
-        client = AlphaVantageClient(api_key="test_key")
-        assert client.api_key == "test_key"
-        assert client.base_url == "https://www.alphavantage.co/query"
-
-        # Test session setup
-        client._setup_sync_session()
-        assert client.session is not None
-
-        # Test close method
-        client.close()
-
-    @patch("ticker_converter.api_clients.client.requests.Session.get")
-    def test_api_client_make_request_success(self, mock_get):
-        """Test successful make_request."""
-        from ticker_converter.api_clients.client import AlphaVantageClient
-
-        # Setup mock response
-        mock_response = Mock()
-        mock_response.status_code = 200
-        mock_response.json.return_value = {
-            "Meta Data": {"1. Information": "Daily Prices"},
-            "Time Series (Daily)": {
-                "2023-01-01": {
-                    "1. open": "100.00",
-                    "2. high": "105.00",
-                    "3. low": "99.00",
-                    "4. close": "104.00",
-                    "5. volume": "1000000",
-                }
-            },
-        }
-        mock_response.raise_for_status.return_value = None
-        mock_get.return_value = mock_response
-
-        client = AlphaVantageClient(api_key="test_key")
-        result = client.make_request(
-            {"function": "TIME_SERIES_DAILY", "symbol": "AAPL"}
-        )
-
-        # Verify result
-        assert "Meta Data" in result
-        assert "Time Series (Daily)" in result
-
-    def test_api_client_daily_stock_data(self):
-        """Test get_daily_stock_data method."""
-        from ticker_converter.api_clients.client import AlphaVantageClient
-
-        with patch.object(AlphaVantageClient, "make_request") as mock_request:
-            mock_request.return_value = {
-                "Meta Data": {"1. Information": "Daily Prices"},
-                "Time Series (Daily)": {
-                    "2023-01-01": {
-                        "1. open": "100.00",
-                        "2. high": "105.00",
-                        "3. low": "99.00",
-                        "4. close": "104.00",
-                        "5. volume": "1000000",
-                    }
-                },
-            }
-
-            client = AlphaVantageClient(api_key="test_key")
-            result = client.get_daily_stock_data("AAPL")
-            assert result is not None
-            mock_request.assert_called_once()
-
-    def test_api_constants_enums(self):
-        """Test API constants and enums."""
-        from ticker_converter.api_clients.constants import (
-            MAX_RETRIES,
-            REQUEST_DELAY_SECONDS,
-            TIMEOUT_SECONDS,
-            AlphaVantageFunction,
-            AlphaVantageResponseKey,
-        )
-
-        # Test enum values
-        assert AlphaVantageFunction.TIME_SERIES_DAILY == "TIME_SERIES_DAILY"
-        assert AlphaVantageFunction.TIME_SERIES_INTRADAY == "TIME_SERIES_INTRADAY"
-        assert AlphaVantageFunction.OVERVIEW == "OVERVIEW"
-
-        # Test constants
-        assert REQUEST_DELAY_SECONDS >= 0
-        assert MAX_RETRIES >= 0
-        assert TIMEOUT_SECONDS > 0
-
-    def test_api_exceptions_hierarchy(self):
-        """Test exception hierarchy."""
-        from ticker_converter.api_clients.exceptions import (
-            AlphaVantageAPIError,
-            AlphaVantageAuthenticationError,
-            AlphaVantageConfigError,
-            AlphaVantageDataError,
-            AlphaVantageRateLimitError,
-            AlphaVantageRequestError,
-            AlphaVantageTimeoutError,
-        )
-
-        # Test base exception
-        base_error = AlphaVantageAPIError("Test error", "E001")
-        assert str(base_error) == "Test error"
-        assert base_error.error_code == "E001"
-
-        # Test specific exceptions inherit from base
-        rate_error = AlphaVantageRateLimitError("Rate limit")
-        assert isinstance(rate_error, AlphaVantageAPIError)
-
-        request_error = AlphaVantageRequestError("Request failed")
-        assert isinstance(request_error, AlphaVantageAPIError)
-
-        config_error = AlphaVantageConfigError("Config error")
-        assert isinstance(config_error, AlphaVantageAPIError)
-
-        auth_error = AlphaVantageAuthenticationError("Auth failed")
-        assert isinstance(auth_error, AlphaVantageAPIError)
-
-        timeout_error = AlphaVantageTimeoutError("Timeout")
-        assert isinstance(timeout_error, AlphaVantageAPIError)
-
-        data_error = AlphaVantageDataError("Data error")
-        assert isinstance(data_error, AlphaVantageAPIError)
-
-    def test_api_utils_functions(self):
-        """Test utility functions that exist."""
-        from ticker_converter.api_clients.utils import (
-            parse_alpha_vantage_date,
-            setup_logging,
-            validate_api_key,
-        )
-
-        # Test logging setup
-        logger = setup_logging("test")
-        assert logger.name == "test"
-
-        # Test API key validation
-        assert validate_api_key("valid_key_123") is True
-        assert validate_api_key("") is False
-        assert validate_api_key(None) is False
-
-        # Test date parsing
-        date_str = parse_alpha_vantage_date("2023-01-01")
-        assert date_str == "2023-01-01"
-
-    def test_config_loading(self):
-        """Test config module loading."""
-        from ticker_converter.config import DATABASE_URL
-
-        # Test that DATABASE_URL exists and is a string
-        assert isinstance(DATABASE_URL, str)
-
-    def test_cli_module(self):
-        """Test CLI module."""
-        from ticker_converter.cli import main
-
-        # Test main function exists
-        assert callable(main)
-
-    def test_data_models_module(self):
-        """Test data models module loads."""
-        import ticker_converter.data_models.api_models as api_models
-        import ticker_converter.data_models.market_data as market_data
-
-        # Test modules load successfully
-        assert hasattr(api_models, "__file__")
-        assert hasattr(market_data, "__file__")
-
-    def test_data_ingestion_modules_import(self):
-        """Test data ingestion modules import successfully."""
-        import ticker_converter.data_ingestion.currency_fetcher as currency
-        import ticker_converter.data_ingestion.database_manager as db_mgr
-        import ticker_converter.data_ingestion.nyse_fetcher as nyse
-
-        # Test modules load
-        assert hasattr(nyse, "__file__")
-        assert hasattr(currency, "__file__")
-        assert hasattr(db_mgr, "__file__")
-
-    def test_nyse_fetcher_class(self):
-        """Test NYSEDataFetcher class exists and can be instantiated."""
-        from ticker_converter.data_ingestion.nyse_fetcher import NYSEDataFetcher
-
-        with patch("ticker_converter.data_ingestion.nyse_fetcher.logger"):
-            fetcher = NYSEDataFetcher()
-            assert fetcher is not None
-
-    def test_currency_fetcher_class(self):
-        """Test CurrencyDataFetcher class exists and can be instantiated."""
-        from ticker_converter.data_ingestion.currency_fetcher import CurrencyDataFetcher
-
-        with patch("ticker_converter.data_ingestion.currency_fetcher.logger"):
-            fetcher = CurrencyDataFetcher()
-            assert fetcher is not None
-
-    def test_database_manager_class(self):
-        """Test DatabaseManager class exists and can be instantiated."""
-        from ticker_converter.data_ingestion.database_manager import DatabaseManager
-
-        # Test with memory database
-        db_manager = DatabaseManager(":memory:")
-        assert db_manager is not None
-
-
-if __name__ == "__main__":
-    unittest.main()
